Order fitted models

Find the models that outperform the others in terms of cross-validation index

ModelOrder(saveWD, plot, zip.file = TRUE, cl = NULL)

Arguments

saveWD
directory where outputs of the cross-validation procedure have been saved or zip.file of this directory. All necessary information have been saved in this directory to be able to compile results. This is the only output of findBestModel in the global R environment.

plot
logical. Either to create the output plots or not. Saved directly in saveWD.

zip.file
TRUE (default) to save all outputs in a zipfile with saveWD, FALSE for no zipfile output, or path to a new zip file

cl
a cluster as made with makeCluster. If empty, nbclust in modelselect_opt will be used.

Value

VeryBestModels_crossV: Best models among all BestForModeltypes: Best models among all in the modeltypes type Figures are produced to visualize ranking and eventually help for a choice

Details

According to the family of model, the index of cross-validation procedure is selected to rank all models fitted for all modeltypes conjointly. Model ranks are then statistically compared. Model with lowest ranks not statistically different than the first ranked model are kept.