Characteristics of the model on which to produce outputs

Characteristics of the model on which to produce outputs

model_select(
  saveWD,
  new.data,
  Num = NULL,
  model = NULL,
  modeltype = NULL,
  powerXI = NULL
)

Arguments

saveWD
directory where outputs of model selection are saved, including AllModels_crossV_ordered.

new.data
SpatialPointsDataFrame dataset used to fit the model.

Num
numeric The 'Num' value of the model retained for further analysis like those saved in AllModels_crossV_ordered. Not useful if model and Family are set. Default to the first model in AllModels_crossV_ordered.

model
character string A user-defined model specification. This requires to set the Family parameter. This overrides effects of 'Num', however, if the model exists with the exact same expression in the list of model fitted, 'Num' will be re-attributed.

modeltype
The model type as chosen among modeltypes available (see modelselect_opt)

powerXI
XI parameter for a tweedie model.

Value

Num, modeltype, Family, formulaX